backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
google-auth==2.23.4
requests==2.31.0
